requirements.txt
colorama==0.4.6
dotenv==0.9.9
greenlet==3.2.0
numpy==2.2.2
pandas==2.2.3
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
scipy==1.15.2
six==1.17.0
SQLAlchemy==2.0.37
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
